+++ schemacrawler-api/src/main/java/schemacrawler/crawl/MutableDatabaseInfo.java
@@ -19,6 +19,7 @@ import java.util.Set;
 import java.util.TreeSet;
 import schemacrawler.schema.DatabaseInfo;
 import schemacrawler.schemacrawler.ModelImplementation;
+import us.fatehi.utility.HostLocation;
 import us.fatehi.utility.database.DatabaseInformation;
 import us.fatehi.utility.property.BaseProductVersion;
 import us.fatehi.utility.property.Property;
@@ -36,6 +37,7 @@ final class MutableDatabaseInfo extends BaseProductVersion implements DatabaseIn
   // Mutable properties collection
   private final Set<Property> serverInfo;
   private final Set<Property> databaseProperties;
+  private HostLocation hostLocation;
 
   public MutableDatabaseInfo(final DatabaseInformation databaseInformation) {
     super(databaseInformation);
@@ -43,6 +45,7 @@ final class MutableDatabaseInfo extends BaseProductVersion implements DatabaseIn
 
     serverInfo = new HashSet<>();
     databaseProperties = new HashSet<>();
+    hostLocation = HostLocation.unknown();
   }
 
   /** {@inheritDoc} */
@@ -56,6 +59,11 @@ final class MutableDatabaseInfo extends BaseProductVersion implements DatabaseIn
   @Override
   public Collection<Property> getServerInfo() {
     return new TreeSet<>(serverInfo);
+  }
+
+  @Override
+  public HostLocation getHostLocation() {
+    return hostLocation;
   }
 
   /** {@inheritDoc} */
@@ -79,6 +87,14 @@ final class MutableDatabaseInfo extends BaseProductVersion implements DatabaseIn
   void addServerInfo(final Property property) {
     if (property != null) {
       serverInfo.add(property);
+    }
+  }
+
+  void setHostLocation(final HostLocation hostLocation) {
+    if (hostLocation != null) {
+      this.hostLocation = hostLocation;
+    } else {
+      this.hostLocation = HostLocation.unknown();
     }
   }
 }