+++ schemacrawler-api/src/main/java/schemacrawler/schemacrawler/SchemaRetrievalOptions.java
@@ -40,6 +40,7 @@ public final class SchemaRetrievalOptions implements Options {
   private final EnumMap<SchemaInfoMetadataRetrievalStrategy, MetadataRetrievalStrategy>
       metadataRetrievalStrategyMap;
   private final Consumer<Connection> connectionInitializer;
+  private final HostLocationExtractor hostLocationExtractor;
 
   protected SchemaRetrievalOptions(final SchemaRetrievalOptionsBuilder builder) {
     final SchemaRetrievalOptionsBuilder bldr =
@@ -56,6 +57,7 @@ public final class SchemaRetrievalOptions implements Options {
     enumDataTypeHelper = bldr.enumDataTypeHelper;
     metadataRetrievalStrategyMap = new EnumMap<>(bldr.metadataRetrievalStrategyMap);
     connectionInitializer = bldr.connectionInitializer;
+    hostLocationExtractor = bldr.hostLocationExtractor;
   }
 
   public MetadataRetrievalStrategy get(
@@ -71,6 +73,10 @@ public final class SchemaRetrievalOptions implements Options {
 
   public DatabaseServerType getDatabaseServerType() {
     return dbServerType;
+  }
+
+  public HostLocationExtractor getHostLocationExtractor() {
+    return hostLocationExtractor;
   }
 
   public EnumDataTypeHelper getEnumDataTypeHelper() {