+++ schemacrawler-api/src/test/java/schemacrawler/test/utility/crawl/LightDatabaseInfo.java
@@ -13,6 +13,7 @@ import java.io.Serializable;
 import java.util.Collection;
 import java.util.List;
 import schemacrawler.schema.DatabaseInfo;
+import us.fatehi.utility.HostLocation;
 import us.fatehi.utility.database.DatabaseInformation;
 import us.fatehi.utility.property.Property;
 
@@ -64,6 +65,11 @@ public class LightDatabaseInfo implements DatabaseInfo {
   @Override
   public Collection<Property> getServerInfo() {
     return List.of();
+  }
+
+  @Override
+  public HostLocation getHostLocation() {
+    return HostLocation.unknown();
   }
 
   @Override