+++ schemacrawler-utility/src/main/java/us/fatehi/utility/HostType.java
@@ -1,0 +1,20 @@
+/*
+ * SchemaCrawler
+ * http://www.schemacrawler.com
+ * Copyright (c) 2000-2026, Sualeh Fatehi <sualeh@hotmail.com>.
+ * All rights reserved.
+ * SPDX-License-Identifier: EPL-2.0
+ */
+
+package us.fatehi.utility;
+
+public enum HostType {
+  unknown,
+  localhost,
+  on_premises,
+  remote_host;
+
+  public String getDescription() {
+    return name().replace('_', ' ');
+  }
+}